pyre-jit-trace: extract FBW walker state to fbw_state.rs
Relocate the 86 fbw_* / capture_fbw_* helpers — feature gates, the
store / append / for-iter journals and rollback, executed-effect and
residual counters, the finish-payload channel, and abort-resume carriers
— from jitcode_dispatch/mod.rs into fbw_state.rs. Pure relocation; the
two thread_local! blocks these functions own (FBW_FINISH_PAYLOAD and the
SELFREC_CA_FOLD_ACTIVE / EXCEPTION_STRING_INLINE_ACTIVE gates) move with
them, the latter two promoted to pub(crate) for the scoped-setter guards
that stay in mod.rs.
